Twitch Stream Monitor
A simple monitor using the twitch API that shows whether or not an
individual channel is currently streaming.

Future updates will include the ability to add and "save" your own
channels. Currently you have to adjust the channel array in index.js.
Future updates will also adjust so that online items are at the top of
the complete list, and solve the small issue with the logos at page
load.

Each channel is displayed as its own div, for styling purposes.